requirements.txt
discord.py
PyNaCl
google_trans_new
itranslate
discord.py[voice]
gTTS